# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeoLocale {
  /** Detected locale code (e.g. 'en', 'es', 'ar', 'zh-Hant') */
  locale: string;
  /**
     * Detected ISO country code, or null if lookup failed
     * @nullable
     */
  country: string | null;
  isRtl: boolean;
}

export type GlobalConfigActiveAdNetwork = typeof GlobalConfigActiveAdNetwork[keyof typeof GlobalConfigActiveAdNetwork];


export const GlobalConfigActiveAdNetwork = {
  admob: 'admob',
  applovin: 'applovin',
  ironsource: 'ironsource',
} as const;

export interface GlobalConfig {
  globalAdEnabled: boolean;
  activeAdNetwork: GlobalConfigActiveAdNetwork;
  maxDailyAdUnlocks: number;
  /** @nullable */
  iosRewardUnitId?: string | null;
  /** @nullable */
  androidRewardUnitId?: string | null;
  reviewModeActive: boolean;
}

export type GlobalConfigUpdateActiveAdNetwork = typeof GlobalConfigUpdateActiveAdNetwork[keyof typeof GlobalConfigUpdateActiveAdNetwork];


export const GlobalConfigUpdateActiveAdNetwork = {
  admob: 'admob',
  applovin: 'applovin',
  ironsource: 'ironsource',
} as const;

export interface GlobalConfigUpdate {
  globalAdEnabled?: boolean;
  activeAdNetwork?: GlobalConfigUpdateActiveAdNetwork;
  maxDailyAdUnlocks?: number;
  iosRewardUnitId?: string;
  androidRewardUnitId?: string;
  reviewModeActive?: boolean;
}

/**
 * Per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn.
 */
export type DramaTitles = {[key: string]: string};

export interface Drama {
  id: string;
  titleEn: string;
  /** Per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn. */
  titles: DramaTitles;
  coverUrl: string;
  /** @nullable */
  description?: string | null;
  tags: string[];
  totalEpisodes: number;
  viewsCount: number;
  freeEpisodesCount: number;
  episodesPerAdUnlock: number;
  interstitialAdFreq: number;
  isPublished: boolean;
  createdAt: string;
}

/**
 * Optional per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn.
 */
export type DramaInputTitles = {[key: string]: string};

export interface DramaInput {
  /** @minLength 1 */
  titleEn: string;
  /** Optional per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn. */
  titles?: DramaInputTitles;
  coverUrl: string;
  description?: string;
  tags?: string[];
  freeEpisodesCount?: number;
  episodesPerAdUnlock?: number;
  interstitialAdFreq?: number;
  isPublished?: boolean;
}

/**
 * Optional per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn.
 */
export type DramaUpdateTitles = {[key: string]: string};

export interface DramaUpdate {
  /** @minLength 1 */
  titleEn?: string;
  /** Optional per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn. */
  titles?: DramaUpdateTitles;
  coverUrl?: string;
  description?: string;
  tags?: string[];
  freeEpisodesCount?: number;
  episodesPerAdUnlock?: number;
  interstitialAdFreq?: number;
  isPublished?: boolean;
}

export interface Episode {
  id: string;
  dramaId: string;
  episodeNumber: number;
  /** @nullable */
  title?: string | null;
  videoUrl: string;
  duration: number;
  createdAt: string;
}

export type DramaWithEpisodes = Drama & {
  episodes: Episode[];
};

export interface EpisodeInput {
  episodeNumber: number;
  title?: string;
  videoUrl: string;
  duration?: number;
}

export interface EpisodeUpdate {
  episodeNumber?: number;
  title?: string;
  videoUrl?: string;
  duration?: number;
}

export interface MonetizationRules {
  freeEpisodesCount: number;
  episodesPerAdUnlock: number;
  interstitialAdFreq: number;
}

export interface PlaybackEpisode {
  episodeNumber: number;
  /** @nullable */
  title?: string | null;
  videoUrl: string;
  isUnlocked: boolean;
}

export interface DramaPlayback {
  dramaId: string;
  title: string;
  coverUrl: string;
  /** @nullable */
  description?: string | null;
  tags: string[];
  monetizationRules: MonetizationRules;
  episodes: PlaybackEpisode[];
}

export interface UnlockRequest {
  userId: string;
  dramaId: string;
  currentEpisodeNumber: number;
}

export interface UnlockResult {
  unlockedEpisodes: number[];
  dailyUnlocksUsed: number;
  dailyUnlocksRemaining: number;
}

export type UserAuthProvider = typeof UserAuthProvider[keyof typeof UserAuthProvider];


export const UserAuthProvider = {
  apple: 'apple',
  google: 'google',
  guest: 'guest',
} as const;

export interface User {
  id: string;
  deviceId: string;
  authProvider: UserAuthProvider;
  createdAt: string;
}

export type UserRegisterInputAuthProvider = typeof UserRegisterInputAuthProvider[keyof typeof UserRegisterInputAuthProvider];


export const UserRegisterInputAuthProvider = {
  apple: 'apple',
  google: 'google',
  guest: 'guest',
} as const;

export interface UserRegisterInput {
  /** @minLength 1 */
  deviceId: string;
  authProvider: UserRegisterInputAuthProvider;
}

export interface DashboardSummary {
  totalDramas: number;
  totalEpisodes: number;
  totalViews: number;
  adUnlocksToday: number;
}

export interface HomeSection {
  id: string;
  name: string;
  sortOrder: number;
  createdAt: string;
}

export interface HomeSectionInput {
  /** @minLength 1 */
  name: string;
  sortOrder?: number;
}

export interface HomeSectionUpdate {
  /** @minLength 1 */
  name?: string;
  sortOrder?: number;
}

export type HomeSectionDramaTitles = {[key: string]: string};

export interface HomeSectionDrama {
  dramaId: string;
  titleEn: string;
  titles?: HomeSectionDramaTitles;
  coverUrl: string;
  sortOrder: number;
}

export type HomeSectionWithDramas = HomeSection & {
  dramas: HomeSectionDrama[];
};

export interface HomeSectionDramaAssignment {
  dramaId: string;
  sortOrder: number;
}

export interface UpdateHomeSectionDramasBody {
  dramas: HomeSectionDramaAssignment[];
}

export interface HomeFeedSection {
  id: string;
  name: string;
  sortOrder: number;
  dramas: Drama[];
}

export type FavoriteDramaTitles = {[key: string]: string};

export interface FavoriteDrama {
  dramaId: string;
  createdAt: string;
  titleEn: string;
  titles?: FavoriteDramaTitles;
  coverUrl: string;
  freeEpisodesCount?: number;
}

export interface WatchProgressEntry {
  dramaId: string;
  lastEpisode: number;
  position: number;
  updatedAt: string;
}

export interface WatchProgressUpdate {
  /** @minimum 1 */
  lastEpisode: number;
  /** @minimum 0 */
  position: number;
}

export type ListDramasParams = {
publishedOnly?: boolean;
search?: string;
};

export type GetDramaPlaybackParams = {
dramaId: string;
userId?: string;
locale?: string;
};