# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dramas:
    get:
      operationId: listDramas
      tags: [dramas]
      summary: List all dramas
      parameters:
        - name: publishedOnly
          in: query
          required: false
          schema:
            type: boolean
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of dramas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Drama"
    post:
      operationId: createDrama
      tags: [dramas]
      summary: Create a drama (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DramaInput"
      responses:
        "201":
          description: Created drama
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drama"
  /dramas/{dramaId}:
    get:
      operationId: getDrama
      tags: [dramas]
      summary: Get a single drama with its episodes (admin edit view)
      parameters:
        - name: dramaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Drama with episodes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DramaWithEpisodes"
        "404":
          description: Not found
    patch:
      operationId: updateDrama
      tags: [dramas]
      summary: Update a drama (admin)
      parameters:
        - name: dramaId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DramaUpdate"
      responses:
        "200":
          description: Updated drama
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Drama"
        "404":
          description: Not found
    delete:
      operationId: deleteDrama
      tags: [dramas]
      summary: Delete a drama (admin)
      parameters:
        - name: dramaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /dramas/playback:
    get:
      operationId: getDramaPlayback
      tags: [dramas]
      summary: Get drama playback details with per-user unlock state
      parameters:
        - name: dramaId
          in: query
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: false
          schema:
            type: string
        - name: locale
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Drama playback details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DramaPlayback"
        "404":
          description: Not found
...
  /users/{userId}/favorites:
    get:
      operationId: listFavorites
      tags: [favorites]
      summary: List a user's favorited dramas
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of favorited dramas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FavoriteDrama"
  /users/{userId}/favorites/{dramaId}:
    put:
      operationId: addFavorite
      tags: [favorites]
      summary: Save a drama to the user's favorites (idempotent)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: dramaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Favorite created or already existed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteDrama"
        "404":
          description: User or drama not found
    delete:
      operationId: removeFavorite
      tags: [favorites]
      summary: Remove a drama from the user's favorites
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: dramaId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed
  /users/{userId}/watch-progress:
    get:
      operationId: listWatchProgress
      tags: [watch-progress]
      summary: List a user's watch progress across all dramas, most recently updated first
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of watch progress entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchProgressEntry"
  /users/{userId}/watch-progress/{dramaId}:
    put:
      operationId: setWatchProgress
      tags: [watch-progress]
      summary: Upsert the user's watch progress for a drama
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: dramaId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchProgressUpdate"
      responses:
        "200":
          description: Updated watch progress entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchProgressEntry"
        "404":
          description: User or drama not found
components:
  schemas:
...
    Drama:
      type: object
      required:
        - id
        - titleEn
        - titles
        - coverUrl
        - tags
        - totalEpisodes
        - viewsCount
        - freeEpisodesCount
        - episodesPerAdUnlock
        - interstitialAdFreq
        - isPublished
        - createdAt
      properties:
        id:
          type: string
        titleEn:
          type: string
        titles:
          type: object
          additionalProperties:
            type: string
          description: "Per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn."
        coverUrl:
          type: string
        description:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        totalEpisodes:
          type: integer
        viewsCount:
          type: integer
        freeEpisodesCount:
          type: integer
        episodesPerAdUnlock:
          type: integer
        interstitialAdFreq:
          type: integer
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DramaInput:
      type: object
      required:
        - titleEn
        - coverUrl
      properties:
        titleEn:
          type: string
          minLength: 1
        titles:
          type: object
          additionalProperties:
            type: string
          description: "Optional per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn."
        coverUrl:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        freeEpisodesCount:
          type: integer
        episodesPerAdUnlock:
          type: integer
        interstitialAdFreq:
          type: integer
        isPublished:
          type: boolean
    DramaUpdate:
      type: object
      properties:
        titleEn:
          type: string
          minLength: 1
        titles:
          type: object
          additionalProperties:
            type: string
          description: "Optional per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn."
        coverUrl:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        freeEpisodesCount:
          type: integer
        episodesPerAdUnlock:
          type: integer
        interstitialAdFreq:
          type: integer
        isPublished:
          type: boolean
    Episode:
      type: object
      required:
        - id
        - dramaId
        - episodeNumber
        - videoUrl
        - duration
        - createdAt
      properties:
        id:
          type: string
        dramaId:
          type: string
        episodeNumber:
          type: integer
        title:
          type: ["string", "null"]
        videoUrl:
          type: string
        duration:
          type: integer
        createdAt:
          type: string
          format: date-time
    DramaWithEpisodes:
      allOf:
        - $ref: "#/components/schemas/Drama"
        - type: object
          required:
            - episodes
          properties:
            episodes:
              type: array
              items:
                $ref: "#/components/schemas/Episode"
...
    MonetizationRules:
      type: object
      required:
        - freeEpisodesCount
        - episodesPerAdUnlock
        - interstitialAdFreq
      properties:
        freeEpisodesCount:
          type: integer
        episodesPerAdUnlock:
          type: integer
        interstitialAdFreq:
          type: integer
    PlaybackEpisode:
      type: object
      required:
        - episodeNumber
        - videoUrl
        - isUnlocked
      properties:
        episodeNumber:
          type: integer
        title:
          type: ["string", "null"]
        videoUrl:
          type: string
        isUnlocked:
          type: boolean
    DramaPlayback:
      type: object
      required:
        - dramaId
        - title
        - coverUrl
        - tags
        - monetizationRules
        - episodes
      properties:
        dramaId:
          type: string
        title:
          type: string
        coverUrl:
          type: string
        description:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        monetizationRules:
          $ref: "#/components/schemas/MonetizationRules"
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/PlaybackEpisode"
...
    FavoriteDrama:
      type: object
      required:
        - dramaId
        - createdAt
        - titleEn
        - coverUrl
      properties:
        dramaId:
          type: string
        createdAt:
          type: string
          format: date-time
        titleEn:
          type: string
        titles:
          type: object
          additionalProperties:
            type: string
        coverUrl:
          type: string
        freeEpisodesCount:
          type: integer
    WatchProgressEntry:
      type: object
      required:
        - dramaId
        - lastEpisode
        - position
        - updatedAt
      properties:
        dramaId:
          type: string
        lastEpisode:
          type: integer
        position:
          type: integer
        updatedAt:
          type: string
          format: date-time
    WatchProgressUpdate:
      type: object
      required:
        - lastEpisode
        - position
      properties:
        lastEpisode:
          type: integer
          minimum: 1
        position:
          type: integer
          minimum: 0