# lib/api-zod/src/generated/types/drama.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DramaTitles } from './dramaTitles';

export interface Drama {
  id: string;
  titleEn: string;
  /** Per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn. */
  titles: DramaTitles;
  coverUrl: string;
  /** @nullable */
  description?: string | null;
  tags: string[];
  totalEpisodes: number;
  viewsCount: number;
  freeEpisodesCount: number;
  episodesPerAdUnlock: number;
  interstitialAdFreq: number;
  isPublished: boolean;
  createdAt: Date;
}