# lib/api-zod/src/generated/types/dramaUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DramaUpdateTitles } from './dramaUpdateTitles';

export interface DramaUpdate {
  /** @minLength 1 */
  titleEn?: string;
  /** Optional per-locale title overrides, keyed by locale code (e.g. 'es', 'th', 'ar'). English falls back to titleEn. */
  titles?: DramaUpdateTitles;
  coverUrl?: string;
  description?: string;
  tags?: string[];
  freeEpisodesCount?: number;
  episodesPerAdUnlock?: number;
  interstitialAdFreq?: number;
  isPublished?: boolean;
}