# lib/api-zod/src/generated/types/favoriteDrama.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FavoriteDramaTitles } from './favoriteDramaTitles';

export interface FavoriteDrama {
  dramaId: string;
  createdAt: Date;
  titleEn: string;
  titles?: FavoriteDramaTitles;
  coverUrl: string;
  freeEpisodesCount?: number;
}